requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
idna==3.6
packaging==23.2
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
